improve version checking functions
