+++ pkg/utilities/version/version.go
@@ -1,6 +1,10 @@
 package version
 
 import (
+	"fmt"
+	"strconv"
+	"strings"
+
 	"github.com/blang/semver"
 	"github.com/pkg/errors"
 )
@@ -38,4 +42,49 @@ func MustMatchRange(version, versionsRange string) bool {
 		panic(err)
 	}
 	return matches
+}
+
+func Jump(nodeVersion, machineVersion string) (bool, error) {
+	nodemajor, nodeminor, _, err := parseVersion(nodeVersion)
+	if err != nil {
+		return false, err
+	}
+	machinemajor, machineminor, _, err := parseVersion(machineVersion)
+	if err != nil {
+		return false, err
+	}
+	return machinemajor == nodemajor && machineminor-nodeminor > 1, nil
+}
+
+func LessThan(v1, v2 string) (bool, error) {
+	v1major, v1minor, v1patch, err := parseVersion(v1)
+	if err != nil {
+		return false, err
+	}
+	v2major, v2minor, v2patch, err := parseVersion(v2)
+	if err != nil {
+		return false, err
+	}
+	return (v1major < v2major) ||
+		(v1major == v2major && v1minor < v2minor) ||
+		(v1major == v2major && v1minor == v2minor && v1patch < v2patch), nil
+}
+
+func parseVersion(v string) (int, int, int, error) {
+	if strings.HasPrefix(v, "v") {
+		v = v[1:]
+	}
+	chunks := strings.Split(v, ".") // drop "v" at front
+	if len(chunks) != 3 {           // major.minor.patch
+		return -1, -1, -1, fmt.Errorf("Invalid kubernetes version: %s", v)
+	}
+	var results = []int{-1, -1, -1}
+	for idx, item := range chunks {
+		val, err := strconv.Atoi(item)
+		if err != nil {
+			return -1, -1, -1, errors.Wrapf(err, "version is invalid")
+		}
+		results[idx] = val
+	}
+	return results[0], results[1], results[2], nil
 }
